Jdempotent-core/src/test/java/br/com/insidesoftwares/jdempotent/core/chain/JdempotentPropertyAnnotationChainTest.java
package br.com.insidesoftwares.jdempotent.core.chain;

import br.com.insidesoftwares.jdempotent.core.model.ChainData;
import br.com.insidesoftwares.jdempotent.core.model.KeyValuePair;
import br.com.insidesoftwares.jdempotent.core.utils.IdempotentTestPayload;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.junit.jupiter.api.Assertions.assertEquals;

@ExtendWith(SpringExtension.class)
class JdempotentPropertyAnnotationChainTest {

    @InjectMocks
    private JdempotentPropertyAnnotationChain jdempotentPropertyAnnotationChain;

    @Test
    void should_process_with_no_annotation() throws IllegalAccessException, NoSuchFieldException {
        //Given
        IdempotentTestPayload idempotentTestPayload = new IdempotentTestPayload();
        idempotentTestPayload.setEventId(1l);
        ChainData chainData = new ChainData();
        chainData.setArgs(idempotentTestPayload);
        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));

        //When
        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);

        //Then
        assertEquals("transactionId", process.getKey());
        assertEquals(1l, process.getValue());
    }

    @Test
    void should_process_with_another_annotated_property() throws IllegalAccessException, NoSuchFieldException {
        //Given
        IdempotentTestPayload idempotentTestPayload = new IdempotentTestPayload();
        idempotentTestPayload.setEventId(1l);
        ChainData chainData = new ChainData();
        chainData.setArgs(idempotentTestPayload);
        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));

        //When
        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);

        //Then
        assertEquals("transactionId", process.getKey());
        assertEquals(1l, process.getValue());
    }
}
